Restructure & Bugfix
# rename playlist font color in dark mode
# restructure helper js files
