+++ app/index.html
@@ -8,7 +8,9 @@
         <link rel="stylesheet" href="./css/content_right.css">
         <!-- <script type="text/javascript" src="./js/rest.js"></script> -->
         <script type="text/javascript" src="./js/icons.js"></script>
-        <script type="text/javascript" src="./js/helper.js"></script>
+        <script type="text/javascript" src="./js/helper_global.js"></script>
+        <script type="text/javascript" src="./js/helper_navigation.js"></script>
+        <script type="text/javascript" src="./js/helper_entries.js"></script>
         <script type="text/javascript" src="./js/request.js"></script>
         <script type="text/javascript" src="./js/drag_handler.js"></script>
         <script type="text/javascript" src="./js/favorite.js"></script>
